organize user files in containers users
